requirements.txt
streamlit
pandas
google-generativeai
requests
python-dotenv
gspread
oauth2client
google-api-python-client
google-auth-oauthlib
google-auth
